fix: 관리자 권한 체크 로직 오류 수정
- 결제 생성/승인/조회 API 통합
- 헥사고날 아키텍처 구조 반영 및 예외 처리 개선

+++ spring_react_payment/toas_payment2_v2/src/main/java/com/sleekydz86/payment2v2/domain/payment/adapter/in/web/dto/PaymentApprovalApiResponse.java
@@ -3,6 +3,8 @@ package com.sleekydz86.payment2v2.domain.payment.adapter.in.web.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class PaymentApprovalApiResponse {
@@ -13,10 +15,9 @@ public class PaymentApprovalApiResponse {
     private String mode;
     private String approvalTime;
     private String stateMsg;
-    private Integer amount;
-    private Integer discountedAmount;
-    private Integer paidAmount;
+    private BigDecimal amount;
+    private BigDecimal discountedAmount;
+    private BigDecimal paidAmount;
     private String payMethod;
     private String transactionId;
 }
-
